Original change: https://android-review.googlesource.com/c/platform/external/adt-infra/+/3012354

+++ pytest/test_embedded/tests/performance/test_pcmark.py
@@ -77,7 +77,8 @@ async def test_pcmark(avd, log_directory, record_property):
         return
 
     temp_path = bundle_path.joinpath("tmp")
-    shutil.rmtree(temp_path)
+    if temp_path.exists():
+        shutil.rmtree(temp_path)
     temp_path.mkdir()
 
     await install_pcmark(avd, bundle_path)
